Employee/src/main/java/com/programmers/apply/service/impl/EmployeeServiceImpl.java
package com.programmers.apply.service.impl;

import java.util.List;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import com.programmers.apply.entities.Employee;
import com.programmers.apply.repository.EmployeeRepository;
import com.programmers.apply.service.EmployeeService;

@Service
public class EmployeeServiceImpl implements EmployeeService {
 
    @Autowired
    private EmployeeRepository employeeRepository;
 
    @Override
    public Employee getEmployeeByName(String name) {
        return employeeRepository.findByName(name);
    }

	@Override
	public List<Employee> getAllEmployees() {
		return employeeRepository.findAll();
	}

	@Override
	public Employee save(Employee employee) {
		return employeeRepository.saveAndFlush(employee);
	}
}
